requirements/base.txt
Django==3.2
django-environ==0.4.5
django-crispy-forms==1.11.1
psycopg2-binary==2.8.6
# Django REST Framework
djangorestframework<4.0  # https://github.com/encode/django-rest-framework
#django-debug-toolbar<1.11  # https://github.com/jazzband/django-debug-toolbar
#line_profiler==3.0.2
#django-debug-toolbar-line-profiler==0.6.1
#openpyxl==2.4.2
#numpy==1.21.2
#pandas==1.3.3
#openpyxl==3.0.9
reportlab==3.6.1
PyPDF2==1.26.0
# QrCode
#qrcode==6.1
django-cleanup==5.2.0
xhtml2pdf==0.2.5